api-03/requirements.txt
apache_libcloud==3.6.0
flasgger
Flask
Flask-Classful
waitress